refactor: Move license info into new module
