+++ src/clientversion.h
@@ -34,11 +34,6 @@ extern const std::string UA_NAME;
 std::string FormatFullVersion();
 std::string FormatSubVersion(const std::string& name, int nClientVersion, const std::vector<std::string>& comments);
 
-std::string CopyrightHolders(const std::string& strPrefix);
-
-/** Returns licensing information (for -version) */
-std::string LicenseInfo();
-
 #endif // RC_INVOKED
 
 #endif // BITCOIN_CLIENTVERSION_H
