+++ src/common/license_info.h
@@ -1,0 +1,15 @@
+// Copyright (c) The Bitcoin Core developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or https://opensource.org/license/mit/.
+
+#ifndef BITCOIN_COMMON_LICENSE_INFO_H
+#define BITCOIN_COMMON_LICENSE_INFO_H
+
+#include <string>
+
+std::string CopyrightHolders(const std::string& strPrefix);
+
+/** Returns licensing information (for -version) */
+std::string LicenseInfo();
+
+#endif // BITCOIN_COMMON_LICENSE_INFO_H
