+++ src/qt/splashscreen.cpp
@@ -7,6 +7,7 @@
 #include <qt/splashscreen.h>
 
 #include <clientversion.h>
+#include <common/license_info.h>
 #include <common/system.h>
 #include <interfaces/handler.h>
 #include <interfaces/node.h>
